+++ lib/fawn.js
@@ -41,7 +41,7 @@ var Fawn = {
     TaskMdl = require("../models/task")(mongoose, collection);
     Task = require("./task")(mongoose, TaskMdl);
     Roller = require("./roller")(mongoose, TaskMdl);
-    utils = require("./utils")(mongoose);
+    utils = require("./utils/db.utils")(mongoose);
   }
 
   , Task: function() {
