+++ lib/roller.js
@@ -6,8 +6,11 @@
  */
 
 var Promise = require("bluebird");
-var constants = require("./constants");
 
+var constants = require("./constants");
+var utils = require("./utils/gen.utils");
+
+// constants
 var INITIAL = constants.INITIAL;
 var DONE = constants.DONE;
 var ROLLED = constants.ROLLED;
@@ -16,10 +19,14 @@ var UPDATE = constants.UPDATE;
 var REMOVE = constants.REMOVE;
 var FILE_REMOVE = constants.FILE_REMOVE;
 var FILE_SAVE = constants.FILE_SAVE;
+
+//gen utils
+var updateState = utils.updateState;
+
+// db utils
 var mongoose;
 var TaskMdl;
-var utils;
-var updateState;
+var dbUtils;
 var getModel;
 
 /**
@@ -34,9 +41,8 @@ var getModel;
 var RollerProvider = function(_mongoose, _TaskMdl) {
   mongoose = _mongoose;
   TaskMdl = _TaskMdl;
-  utils = require("./utils")(mongoose);
-  updateState = utils.updateState;
-  getModel = utils.getModel;
+  dbUtils = require("./utils/db.utils")(mongoose);
+  getModel = dbUtils.getModel;
 
   return Roller;
 };
@@ -177,7 +183,7 @@ function rollbackRemoveOrUpdate(step, task) {
 function rollbackFileSave(step, task) {
   var gfs = new Grid(mongoose.connection.db);
 
-  return utils.removeFile(step.dataStore[0]._id, gfs)
+  return dbUtils.removeFile(step.dataStore[0]._id, gfs)
     .then(function () {
       return updateState(task, step.index, ROLLED);
     });
@@ -204,7 +210,7 @@ function rollbackFileRemove(step, task) {
         if (!shadowFile) return resolve();
 
         function done() {
-          utils.removeFile(data.shadow, gfs).then(function () {
+          dbUtils.removeFile(data.shadow, gfs).then(function () {
             updateState(task, step.index, ROLLED)
               .then(resolve)
               .catch(reject);
