+++ lib/task.js
@@ -5,6 +5,17 @@ var Grid = require("gridfs-stream");
 var fs = require("fs");
 
 var constants = require("./constants");
+var utils = require("./utils/gen.utils");
+
+// general utility functions
+var validModel = utils.validModel;
+var validDoc = utils.validDoc;
+var getModelName = utils.getModelName;
+var xcode = utils.xcode;
+var resolveFuture = utils.resolveFuture;
+var isObject = utils.isObject;
+var storeOldData = utils.storeOldData;
+var updateState = utils.updateState;
 
 // didn't want to type "constants." every time
 var INITIAL = constants.INITIAL;
@@ -18,14 +29,12 @@ var FILE_REMOVE = constants.FILE_REMOVE;
 
 // variables which require a mongoose instance
 var mongoose;
-var utils;
-var updateState;
+var dbUtils;
 var modelCache;
 var setModel;
 var getModel;
 var TaskMdl;
 var Roller;
-var xcode;
 
 /**
  * Provider for the Task class. It initializes all the
@@ -43,12 +52,10 @@ var TaskProvider = function (_mongoose, _TaskMdl) {
   Grid.mongo = mongoose.mongo;
   TaskMdl = _TaskMdl;
   Roller = require("./roller")(mongoose, _TaskMdl);
-  utils = require("./utils")(mongoose);
-  updateState = utils.updateState;
-  modelCache = utils.modelCache;
-  setModel = utils.setModel;
-  getModel = utils.getModel;
-  xcode = utils.xcode;
+  dbUtils = require("./utils/db.utils")(mongoose);
+  modelCache = dbUtils.modelCache;
+  setModel = dbUtils.setModel;
+  getModel = dbUtils.getModel;
 
   return Task;
 };
@@ -77,14 +84,14 @@ var Task = function() {
   };
 
   /**
-   * @see utils.initModel
+   * @see dbUtils.initModel
    *
    * @param modelName The intended name of the model
    * @param schema The schema associated with this model
    * @returns {Task}
    */
   task.initModel = function(modelName, schema) {
-    utils.initModel(modelName, schema);
+    dbUtils.initModel(modelName, schema);
 
     return task;
   };
@@ -293,7 +300,7 @@ var Task = function() {
 
       return chain.then(function (results) {
         var gfs = Grid(mongoose.connection.db);
-        var removeChain = utils.makeRemoveChain(_task, gfs);
+        var removeChain = dbUtils.makeRemoveChain(_task, gfs);
 
         return removeChain.then(function () {
           return _task.remove()
@@ -341,11 +348,12 @@ function performUpdate(step, task, results) {
   var Collection = getModel(step.name);
   var condition = xcode(step.condition);
   var data = xcode(step.data);
+  var db = mongoose.connection.db;
 
   resolveFuture(condition, results);
   resolveFuture(data, results);
 
-  return storeOldData(step, condition).then(function () {
+  return storeOldData(db, step, condition).then(function () {
     return updateState(task, step.index, PENDING).then(function () {
       return Collection.update(condition, data, step.options)
         .exec()
@@ -376,7 +384,7 @@ function performSave(step, task, results) {
   var doc = new Collection(data);
   var dataStore = [];
 
-  doc._id = doc._id || utils.generateId();
+  doc._id = doc._id || dbUtils.generateId();
   dataStore.push({_id: doc._id});
 
   step.dataStore = dataStore;
@@ -405,10 +413,11 @@ function performSave(step, task, results) {
 function performRemove(step, task, results) {
   var Collection = getModel(step.name);
   var condition = xcode(step.condition);
+  var db = mongoose.connection.db;
 
   resolveFuture(condition, results);
 
-  return storeOldData(step, condition).then(function () {
+  return storeOldData(db, step, condition).then(function () {
     return updateState(task, step.index, PENDING).then(function () {
       return Collection.remove(condition).exec().then(function (result) {
         results.push(result);
@@ -434,7 +443,7 @@ function performFileSave(step, task, results) {
 
   resolveFuture(options, results);
 
-  options._id = options._id || utils.generateId();
+  options._id = options._id || dbUtils.generateId();
   dataStore.push({_id: options._id});
 
   step.options = options;
@@ -489,7 +498,7 @@ function performFileRemove(step, task, results) {
       else{
         var gfs = Grid(mongoose.connection.db);
 
-        chain = utils.removeFile(options, gfs).then(function (result) {
+        chain = dbUtils.removeFile(options, gfs).then(function (result) {
           results.push(result);
         });
       }
@@ -519,7 +528,7 @@ function storeOldFile(step) {
       if (err) return reject(err);
       if (!file) return resolve(false);
 
-      dataStore.push({removed: file._id, shadow: utils.generateId()});
+      dataStore.push({removed: file._id, shadow: dbUtils.generateId()});
       step.dataStore = dataStore;
 
       step.save().then(function () {
@@ -531,129 +540,6 @@ function storeOldFile(step) {
         gfs.createReadStream({_id: file._id}).pipe(writeStream);
       });
     });
-  });
-}
-
-/**
- * This function stores data that's about to be
- * changed by a step, for rollback purposes
- *
- * @param step the step
- * @param condition literal obj rep of the step's condition
- *
- * @returns {Promise|*}
- */
-function storeOldData(step, condition){
-  var db = mongoose.connection.db;
-  var Collection = db.collection(step.name);
-  var options = step.options
-    ? xcode(step.options)
-    : step.type === constants.REMOVE ? {multi: true} : null;
-  var query = Collection.find(condition);
-  var searchQuery = options && options.multi === true
-    ? query
-    : query.limit(1);
-
-  return searchQuery.toArray().then(function(result){
-    step.dataStore = result;
-  });
-}
-
-/**
- * Loosely check that a mongoose model is valid
- *
- * @param model the model to check
- *
- * @returns {boolean}
- */
-function validModel(model){
-  if (typeof model === "string") return true;
-
-  return typeof model === "function" && model.modelName;
-}
-
-/**
- * Loosely checks that a mongoose document is valid.
- *
- * @param doc the document to check
- *
- * @returns {boolean}
- */
-function validDoc(doc){
-  return doc && doc._doc && doc.constructor && doc.constructor.modelName;
-}
-
-/**
- * Checks that an object is valid
- *    isObject({name: "Max Payne"}) = true
- *    isObject(new Array()) = false
- *    isObject(null) = false
- *
- * @param obj the object to validate
- *
- * @returns {boolean}
- */
-function isObject(obj) {
-  return obj && obj.constructor && obj.constructor === Object;
-}
-
-/**
- * Gets the name of a model
- *
- * @param model the model
- *
- * @returns the model's name
- */
-function getModelName(model) {
-  return typeof model === "string" ? model : model.modelName;
-}
-
-/**
- * Replaces a placeholder with real data.
- *
- * e.g:
- *  var testResults = [{name: "Bob"}]
- *  var testObj = {name: {$ojFuture: "0.name"}}
- *
- *  resolveFuture(testObj, testResults);
- *
- *  console.log(testObj) // {name: "Bob"}
- *
- * @param obj Object that might contain templated data
- * @param results results used to replace templated data
- */
-function resolveFuture(obj, results) {
-  var ojf = constants.OJ_FUTURE;
-
-  Object.keys(obj).forEach(function (key) {
-    if (!isObject(obj[key])) return;
-    if (!obj[key][ojf]) return resolveFuture(obj[key], results);
-
-    var placeholder = obj[key][ojf];
-
-    // format: "index.property.property.-----.property"
-    var parts = placeholder.split(".");
-    if (isNaN(parts[0])) throw new Error("step index must be a number");
-
-    var index = parseInt(parts[0]);
-    if (index >= results.length) throw new Error("Invalid index. The result at that index does not exist (yet)");
-
-    var result = results[index];
-
-    for (var i = 1; i < parts.length; i++) {
-      if (result.toObject) result = result.toObject();
-
-      if (result instanceof Array) {
-        if (isNaN(parts[i])) throw new Error("Array index must be a number");
-        parts[i] = parseInt(parts[i]);
-      }
-      else if (!isObject(result))throw new Error("Can't use keys on non-object");
-
-      if (!result[parts[i]]) throw new Error("No such key exists in result");
-      result = result[parts[i]];
-    }
-
-    obj[key] = result;
   });
 }
