+++ lib/utils/db.utils.js
@@ -7,33 +7,11 @@
 var modelCache = {};
 var Promise = require("bluebird");
 
-var constants = require("./constants");
+var constants = require("../constants");
 
 module.exports = function(_mongoose) {
   var mongoose = _mongoose || require("mongoose");
   var Schema = mongoose.Schema;
-
-  /**
-   * Updates the state of a task's step.
-   *
-   * @param task the task
-   * @param index the index of the step to update
-   * @param state the new state of the step
-   * @param results array of results from previous steps
-   *
-   * @returns {Promise|*}
-   */
-  function updateState(task, index, state, results){
-    var TskMdl = task.constructor;
-    var taskObj;
-
-    task.steps[index].state = state;
-    taskObj = task.toObject();
-
-    return TskMdl.collection.updateOne({_id: task._id}, taskObj).then(function(){
-      return Promise.resolve(results);
-    });
-  }
 
   /**
    * gets a collection as a mongoose model.
@@ -187,21 +165,8 @@ module.exports = function(_mongoose) {
     return mongoose.Types.ObjectId();
   }
 
-  /**
-   * For encoding user provided conditions, options, and data as
-   * JSON strings for writes and decoding JSON strings to objects
-   * before operations.
-   *
-   * @param item a JSON string or an object
-   * @returns object or JSON string
-   */
-  function xcode(item) {
-    return typeof item === "string" ? JSON.parse(item) : JSON.stringify(item);
-  }
-
   return {
-    updateState: updateState
-    , setModel: setModel
+    setModel: setModel
     , getModel: getModel
     , initModel: initModel
     , modelCache: modelCache
@@ -210,6 +175,5 @@ module.exports = function(_mongoose) {
     , removeFile: removeFile
     , fileExists: fileExists
     , makeRemoveChain: makeRemoveChain
-    , xcode: xcode
   };
 };
